% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koladaAPI.R
\name{get_api_data}
\alias{get_api_data}
\title{Gets data from the API}
\usage{
get_api_data(endpoint, entity, param)
}
\arguments{
\item{endpoint}{The API endpoint URL}

\item{params}{A list of query parameters}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    Parsed API response
}\if{html}{\out{</div>}}
}
\description{
Gets data from the API
}
