% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getquerydata.R
\name{get_query_data}
\alias{get_query_data}
\title{Gets multiple data from the API}
\usage{
get_query_data(endpoint, data_api, entities)
}
\arguments{
\item{endpoint}{The API endpoint URL}

\item{data_api}{The type of data description}

\item{entities}{A list of query parameters}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    Parsed API response
}\if{html}{\out{</div>}}
}
\description{
Gets multiple data from the API
}
