% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getquerydata.R
\name{getquerydata}
\alias{getquerydata}
\title{Gets multiple data from the API}
\usage{
getquerydata(endpoint, entity, municipality, year)
}
\arguments{
\item{endpoint}{The API endpoint URL}

\item{entity}{A list of query parameters}

\item{municipality}{A variable for the municipalities in sweden}

\item{year}{A period of years ranging from 2010 to 2022}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    Parsed API response
}\if{html}{\out{</div>}}
}
\description{
Gets multiple data from the API
}
