% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipalities.R
\docType{data}
\name{municipalities}
\alias{municipalities}
\title{Municipalities data from Statistics Sweden that shows all municipalities in numerical order.}
\format{
\subsection{\code{municipalities}}{

A data frame with 290 rows and 2 columns:
\describe{
\item{Municipality}{names of all municipalities in Sweden}
\item{ID}{Corresponding Municipality IDs}

...
}
}
}
\source{
\url{https://www.scb.se/en/finding-statistics/regional-statistics/regional-divisions/counties-and-municipalities/counties-and-municipalities-in-numerical-order/}
}
\usage{
municipalities
}
\description{
Municipalities data from Statistics Sweden that shows all municipalities in numerical order.
}
\keyword{datasets}
