% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getapidata.R
\name{getapidata}
\alias{getapidata}
\title{Gets single data from the API}
\usage{
getapidata(endpoint, entity, params = list())
}
\arguments{
\item{endpoint}{The API endpoint URL}

\item{params}{A list of query parameters}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    Parsed API response
}\if{html}{\out{</div>}}
}
\description{
Gets single data from the API
}
\examples{
api_data <- getapidata("http://api.kolada.se/v2", entity = "municipality", params = list())
api_data <- getapidata("http://api.kolada.se/v2", "kpi", list())


}
